Version 0.2
Default is changed to not use syslog and not use dotlock.
Add option to use a dotlock file locking
Add option to log to syslog as well as stderr.
Use mail.info facility for syslog.
Support newer fcntl locking as well as older flock.
Check for a non-stale .lock file always.
Improve documentation.